Now casting some values to float in super() call within constructor.
git-svn-id: http://chianti.ucsd.edu/svn@3588 0ecc0d97-ab19-0410-9704-bfe1a75892f5

+++ csplugins/trunk/ucsd/nlandys/fung/src/cytoscape/fung/NodeLabel.java
@@ -19,7 +19,7 @@ public final class NodeLabel extends Label
                    final byte justify)
   {
     super(text, font, scaleFactor, paint, textAnchor,
-          offsetVector.getX(), offsetVector.getY(), justify);
+          (float) offsetVector.getX(), (float) offsetVector.getY(), justify);
     m_nodeAnchor = nodeAnchor;
     switch (m_nodeAnchor) {
     case ANCHOR_CENTER:
